+++ src/main/java/KtlintConfigForm.kt
@@ -4,6 +4,7 @@ import com.nbadal.ktlint.KtlintConfigStorage
 import javax.swing.JCheckBox
 import javax.swing.JComponent
 import javax.swing.JPanel
+import javax.swing.JTextField
 
 class KtlintConfigForm(private val config: KtlintConfigStorage) {
 
@@ -11,6 +12,7 @@ class KtlintConfigForm(private val config: KtlintConfigStorage) {
     private lateinit var enableKtlint: JCheckBox
     private lateinit var enableExperimental: JCheckBox
     private lateinit var treatAsErrors: JCheckBox
+    private lateinit var disabledRules: JTextField
 
     fun createComponent(): JComponent {
         mainPanel.border = IdeBorderFactory.createTitledBorder("Ktlint Settings")
@@ -28,18 +30,21 @@ class KtlintConfigForm(private val config: KtlintConfigStorage) {
         config.enableKtlint = enableKtlint.isSelected
         config.useExperimental = enableExperimental.isSelected
         config.treatAsErrors = treatAsErrors.isSelected
+        config.disabledRules = disabledRules.text
     }
 
     fun reset() {
         enableKtlint.isSelected = config.enableKtlint
         enableExperimental.isSelected = config.useExperimental
         treatAsErrors.isSelected = config.treatAsErrors
+        disabledRules.text = config.disabledRules
     }
 
     val isModified
         get() = !(
             Comparing.equal(config.enableKtlint, enableKtlint.isSelected) &&
                 Comparing.equal(config.useExperimental, enableExperimental.isSelected) &&
-                Comparing.equal(config.treatAsErrors, treatAsErrors.isSelected)
+                Comparing.equal(config.treatAsErrors, treatAsErrors.isSelected) &&
+                Comparing.equal(config.disabledRules, disabledRules.text)
             )
 }